requirements.txt
streamlit>=1.25.0
google-generativeai>=0.2.0
python-dotenv>=1.0.0
SpeechRecognition>=3.9.0
sounddevice>=0.4.9
soundfile>=0.12.1
gTTS>=2.3.2
python-docx>=0.8.12
PyPDF2>=3.0.0,<3.1.0
fpdf>=1.7.2
